+++ packages/inngest/src/components/connect/types.ts
@@ -1,6 +1,8 @@
 import { type RegisterOptions } from "../../types.js";
 import { type InngestFunction } from "../InngestFunction.js";
 
+export const DEFAULT_SHUTDOWN_SIGNALS = ["SIGINT", "SIGTERM"];
+
 export interface ConnectHandlerOptions extends RegisterOptions {
   /**
    * An array of the functions to serve and register with Inngest.
@@ -9,7 +11,12 @@ export interface ConnectHandlerOptions extends RegisterOptions {
 
   instanceId: string;
   maxConcurrency?: number;
-  abortSignal?: AbortSignal;
+
+  /**
+   * By default, connections will be gracefully shut down when the current
+   * process receives a SIGINT or SIGTERM signal. Set this to an empty array to disable this behavior.
+   */
+  handleShutdownSignals?: string[];
 }
 
 export interface WorkerConnection {
